requirements.txt
requests 
flask 
flask-sqlalchemy
flask-dance
oauthlib
python-dotenv
google-genai
